toolkit/components/aboutmemory/tests/test_aboutmemory3.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<window title="about:memory"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>

  <!-- This file tests the saving and loading of memory reports to/from file in
       about:memory. -->

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml"></body>

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[
  "use strict";

  const Cc = Components.classes;
  const Ci = Components.interfaces;
  let mgr = Cc["@mozilla.org/memory-reporter-manager;1"].
            getService(Ci.nsIMemoryReporterManager);

  // Remove all the real reporters and multi-reporters;  save them to
  // restore at the end.
  mgr.blockRegistration();
  let e = mgr.enumerateReporters();
  let realReporters = [];
  while (e.hasMoreElements()) {
    let r = e.getNext().QueryInterface(Ci.nsIMemoryReporter);
    mgr.unregisterReporter(r);
    realReporters.push(r);
  }
  e = mgr.enumerateMultiReporters();
  let realMultiReporters = [];
  while (e.hasMoreElements()) {
    let r = e.getNext().QueryInterface(Ci.nsIMemoryMultiReporter);
    mgr.unregisterMultiReporter(r);
    realMultiReporters.push(r);
  }

  // Setup a minimal number of fake reporters.
  const KB = 1024;
  const MB = KB * KB;
  const HEAP  = Ci.nsIMemoryReporter.KIND_HEAP;
  const OTHER = Ci.nsIMemoryReporter.KIND_OTHER;
  const BYTES = Ci.nsIMemoryReporter.UNITS_BYTES;

  function f(aPath, aKind, aAmount, aDesc) {
    return {
      process:     "",
      path:        aPath,
      kind:        aKind,
      units:       BYTES,
      amount:      aAmount,
      description: aDesc
    };
  }

  let fakeReporters = [
    f("heap-allocated",     OTHER,   250 * MB, "Heap allocated."),
    f("explicit/a/b",       HEAP,     50 * MB, "A b."),
    f("other/a",            OTHER,   0.2 * MB, "Other a."),
    f("other/b",            OTHER,   0.1 * MB, "Other b."),
  ];

  for (let i = 0; i < fakeReporters.length; i++) {
    mgr.registerReporterEvenIfBlocked(fakeReporters[i]);
  }

  ]]>
  </script>

  <iframe id="amGoodFrame"  height="200" src="about:memory"></iframe>
  <iframe id="amGoodFrame2" height="200" src="about:memory"></iframe>
  <iframe id="amBadFrame"   height="200" src="about:memory"></iframe>

  <script type="application/javascript">
  <![CDATA[
  function finish()
  {
    // Unregister fake reporters and multi-reporters, re-register the real
    // reporters and multi-reporters, just in case subsequent tests rely on
    // them.
    for (let i = 0; i < fakeReporters.length; i++) {
      mgr.unregisterReporter(fakeReporters[i]);
    }
    for (let i = 0; i < realReporters.length; i++) {
      mgr.registerReporterEvenIfBlocked(realReporters[i]);
    }
    for (let i = 0; i < realMultiReporters.length; i++) {
      mgr.registerMultiReporterEvenIfBlocked(realMultiReporters[i]);
    }
    mgr.unblockRegistration();

    SimpleTest.finish();
  }

  // Load the given file into the frame, then copy+paste the entire frame and
  // check that the cut text matches what we expect.
  function test(aFrameId, aFilename, aExpected, aDumpFirst, aNext) {
    let frame = document.getElementById(aFrameId);
    frame.focus();

    let file = Cc["@mozilla.org/file/directory_service;1"]
               .getService(Components.interfaces.nsIProperties)
               .get("CurWorkD", Components.interfaces.nsIFile);
    file.append("chrome");
    file.append("toolkit");
    file.append("components");
    file.append("aboutmemory");
    file.append("tests");
    file.append(aFilename);

    if (aDumpFirst) {
        let dumper = Cc["@mozilla.org/memory-info-dumper;1"].
                        getService(Ci.nsIMemoryInfoDumper);

        dumper.dumpMemoryReportsToNamedFile(file.path,
                                            /* minimizeMemoryUsage = */ false,
                                            /* dumpChildProcesses = */ false);
    }

    let input = frame.contentWindow.document.getElementById("filePickerInput");
    input.value = file.path;    // this works because it's a chrome test

    var e = document.createEvent('Event');
    e.initEvent('change', true, true);
    input.dispatchEvent(e);

    // Initialize the clipboard contents.
    SpecialPowers.clipboardCopyString("initial clipboard value");

    let numFailures = 0, maxFailures = 30;

    // Because the file load is async, we don't know when it will finish and
    // the output will show up.  So we poll.
    function copyPasteAndCheck() {
      // Copy and paste frame contents, and filter out non-deterministic
      // differences.
      synthesizeKey("A", {accelKey: true});
      synthesizeKey("C", {accelKey: true});
      let actual = SpecialPowers.getClipboardData("text/unicode");
      actual = actual.replace(/\(pid \d+\)/, "(pid NNN)");

      if (actual === aExpected) {
        SimpleTest.ok(true, "Clipboard has the expected contents");
        aNext();
      } else {
        numFailures++;
        if (numFailures === maxFailures) {
          ok(false, "pasted text doesn't match");
          dump("******EXPECTED******\n");
          dump(aExpected);
          dump("*******ACTUAL*******\n");
          dump(actual);
          dump("********************\n");
          finish();
        } else {
          setTimeout(copyPasteAndCheck, 100);
        }
      }
    }
    copyPasteAndCheck();
  }

  // Returns a function that chains together multiple test() calls.
  function chain(aFrameIds) {
    let x = aFrameIds.shift();
    if (x) {
      return function() { test(x.frameId, x.filename, x.expected, x.dumpFirst, chain(aFrameIds)); }
    } else {
      return function() { finish(); };
    }
  }

  // This is pretty simple output, but that's ok;  this file is about testing
  // the loading of data from file.  If we got this far, we're doing fine.
  let expectedGood =
"\
Main Process (pid NNN)\n\
Explicit Allocations\n\
\n\
250.00 MB (100.0%) -- explicit\n\
├──200.00 MB (80.00%) ── heap-unclassified\n\
└───50.00 MB (20.00%) ── a/b\n\
\n\
Other Measurements\n\
\n\
0.30 MB (100.0%) -- other\n\
├──0.20 MB (66.67%) ── a\n\
└──0.10 MB (33.33%) ── b\n\
\n\
250.00 MB ── heap-allocated\n\
\n\
";

  // This is the output for a malformed data file.
  let expectedBad =
"\
Invalid memory report(s): missing 'hasMozMallocUsableSize' property";

  let frames = [
    // This loads a pre-existing file that is valid.
    { frameId: "amGoodFrame", filename: "memory-reports-good.json", expected: expectedGood, dumpFirst: false },

    // This dumps to a file and then reads it back in.  The output is the same as the first test.
    { frameId: "amGoodFrame2", filename: "memory-reports-dumped.json.gz", expected: expectedGood, dumpFirst: true },

    // This loads a pre-existing file that is invalid.
    { frameId: "amBadFrame",  filename: "memory-reports-bad.json",  expected: expectedBad, dumpFirst: false }
  ];

  SimpleTest.waitForFocus(chain(frames));

  SimpleTest.waitForExplicitFinish();
  ]]>
  </script>
</window>
